==> 19_stacking.ipynb <==
from pdb import set_trace

import numpy as np
import pandas as pd
from sklearn.model_selection import StratifiedKFold
from sklearn.linear_model import LogisticRegression
from tqdm import tqdm_notebook as tqdm

import torch
import torch.nn as nn
import torch.nn.functional as F
from torch.utils.data import DataLoader, TensorDataset

from basedir import ROOT

f1 = pd.read_csv('stacked/train/densenet121_long_training.csv')
f2 = pd.read_csv('stacked/train/densenet121_with_tricks.csv')
f3 = pd.read_csv('stacked/train/densenet169.csv')

trn_df = pd.read_csv(ROOT/'train.csv')

X = np.column_stack([f.drop(columns='id_code').values for f in (f1, f2, f3)])
y = trn_df['sirna'].values

X.shape, y.shape

X = torch.tensor(X).float()
y = torch.tensor(y)
print(X.shape, X.type(), y.shape, y.type())

dataset = TensorDataset(X, y)

class LogisticClassifier(nn.Module):
    def __init__(self, n_classes):
        super().__init__()
        self.fc1 = nn.Linear(3324, 2048)
        self.fc2 = nn.Linear(2048, 1024)
        self.logreg = nn.Linear(1024, n_classes)
        
    def forward(self, x):
        x = F.relu(self.fc1(x))
        x = F.relu(self.fc2(x))
        x = self.logreg(x)
        return x

device = torch.device('cuda:0')
logreg = LogisticClassifier(1108)
logreg.to(device)
opt = torch.optim.AdamW(logreg.parameters())

loss_fn = nn.CrossEntropyLoss()

for i in range(10):
    for j, (xb, yb) in enumerate(DataLoader(dataset, shuffle=True, batch_size=1024)):
        opt.zero_grad()
        xb, yb = xb.to(device), yb.to(device)
        out = logreg(xb)
        loss = loss_fn(out, yb)
        loss.backward()
        opt.step()
        print(f'[{i}][{j}] Loss: {loss.item():.4f}')

f1 = pd.read_csv('stacked/test/densenet121_long_training.csv')
f2 = pd.read_csv('stacked/test/densenet121_with_tricks.csv')
f3 = pd.read_csv('stacked/test/densenet169.csv')
X_test = np.column_stack([f.drop(columns='id_code').values for f in (f1, f2, f3)])

X_test = torch.tensor(X_test).to(device).float()

with torch.no_grad():
    logreg.eval()
    probs = []
    for xb in tqdm(DataLoader(X_test, batch_size=128, shuffle=False)):
        out = logreg(xb)
        out = out.softmax(dim=-1).cpu().numpy()
        probs.extend(out)

stacked.shape

stacked = np.row_stack(preds)

sub = pd.DataFrame({'id_code': f1.id_code, 'sirna': stacked.argmax(axis=-1)})

trn_csv = pd.read_csv(ROOT/'train.csv')
tst_csv = pd.read_csv(ROOT/'test.csv')

plate_groups = np.zeros((1108,4), int)
for sirna in range(1108):
    grp = trn_csv.loc[trn_csv.sirna==sirna,:].plate.value_counts().index.values
    assert len(grp) == 3
    plate_groups[sirna,0:3] = grp
    plate_groups[sirna,3] = 10 - grp.sum()

all_test_exp = tst_csv.experiment.unique()

group_plate_probs = np.zeros((len(all_test_exp),4))
for idx in range(len(all_test_exp)):
    preds = sub.loc[tst_csv.experiment == all_test_exp[idx],'sirna'].values
    pp_mult = np.zeros((len(preds),1108))
    pp_mult[range(len(preds)),preds] = 1
    
    sub_test = tst_csv.loc[tst_csv.experiment == all_test_exp[idx],:]
    assert len(pp_mult) == len(sub_test)
    
    for j in range(4):
        mask = np.repeat(plate_groups[np.newaxis, :, j], len(pp_mult), axis=0) == \
               np.repeat(sub_test.plate.values[:, np.newaxis], 1108, axis=1)
        
        group_plate_probs[idx,j] = np.array(pp_mult)[mask].sum()/len(pp_mult)

pd.DataFrame(group_plate_probs, index=all_test_exp)

exp_to_group = group_plate_probs.argmax(1)
print(exp_to_group)

def select_plate_group(pp_mult, idx):
    sub_test = tst_csv.loc[tst_csv.experiment == all_test_exp[idx],:]
    assert len(pp_mult) == len(sub_test)
    mask = np.repeat(plate_groups[np.newaxis, :, exp_to_group[idx]], len(pp_mult), axis=0) != \
           np.repeat(sub_test.plate.values[:, np.newaxis], 1108, axis=1)
    pp_mult[mask] = 0
    return pp_mult

sub = sub.set_index('id_code')

for idx in range(len(all_test_exp)):
    indexes = tst_csv.experiment == all_test_exp[idx]
    preds = stacked[indexes, :].copy()
    preds = select_plate_group(preds, idx)
    sub.loc[tst_csv.id_code[indexes], 'sirna'] = preds.argmax(1)

sub = sub.reset_index()

from IPython.display import FileLink
sub.to_csv('stack_leak.csv', index=False, columns=['id_code', 'sirna'])
FileLink('stack_leak.csv')
